backend/requirements.txt
fastapi
uvicorn
pyodbc
python-dotenv
pandas
openai>=1.0.0
